requirements.txt
fastapi
uvicorn
starlette
pydantic
SQLAlchemy
mysql-connector-python
python-dotenv
